+++ src/utils.js
@@ -1,3 +1,5 @@
+const JSON5 = require('json5');
+
 /**
  * Converts the win path to POSIX standard.
  * The require() function understands only POSIX format.
@@ -10,15 +12,6 @@
  * @return {*}
  */
 const pathToPosix = (value) => value.replace(/\\/g, '/');
-
-/**
- * @param {string} request
- * @return {{resource: string, query: string|null}}
- */
-const parseRequest = (request) => {
-  const [resource, query] = request.split('?');
-  return { resource, query };
-};
 
 const isFunction = (value) => typeof value === 'function';
 
@@ -44,7 +37,68 @@ const shallowEqual = function (obj1, obj2) {
   return true;
 };
 
+/**
+ * Parse resource path and raw query from request.
+ *
+ * @param {string} request
+ * @return {{resource: string, query: string|null}}
+ */
+const parseRequest = (request) => {
+  const [resource, query] = request.split('?');
+  return { resource, query };
+};
+
+/**
+ * Parse request query.
+ *
+ * @param {string} request
+ * @returns {{}}
+ */
+const parseQuery = (request) => {
+  const [, query] = request.split('?');
+  if (!query) return {};
+
+  if (query[0] === '{' && query.slice(-1) === '}') {
+    // TODO: write own micro parser to avoid external dependency of json5 module
+    return JSON5.parse(decodeURIComponent(query));
+  }
+
+  const specialValues = {
+    null: null,
+    true: true,
+    false: false,
+  };
+  const queryArgs = query.split(/[,&]/g);
+  const result = {};
+
+  for (let arg of queryArgs) {
+    let [name, value] = arg.split('=');
+
+    if (value) {
+      value = decodeURIComponent(value);
+
+      if (specialValues.hasOwnProperty(value)) {
+        value = specialValues[value];
+      }
+
+      if (name.slice(-2) === '[]') {
+        name = decodeURIComponent(name.slice(0, -2));
+        if (!Array.isArray(result[name])) {
+          result[name] = [];
+        }
+        result[name].push(value);
+      } else {
+        name = decodeURIComponent(name);
+        result[name] = value;
+      }
+    }
+  }
+
+  return result;
+};
+
 module.exports = {
+  parseQuery,
   pathToPosix,
   parseRequest,
   isFunction,