+++ test/cases/require-responsive-images/webpack.config.js
@@ -52,9 +52,7 @@ module.exports = {
             // note: github generate different hash as local test
             //name: 'assets/img/[name].[hash:8]-[width]w.[ext]',
             name: 'assets/img/[name]-[width]w.[ext]',
-            sizes: [120, 200, 320],
-            format: 'webp',
-            //placeholder: true, // NOTE: v3.0.2 has bug, this option must be false
+            sizes: [640], // default size for all images
           },
         },
       },