+++ test/manual/require-responsive-image/src/js/main.js
@@ -1,2 +1,0 @@
-console.log('\nHello PUG PLUGIN!');
-console.log('\n');